requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
moviepy==1.0.3
Pillow==10.1.0
urllib3==1.26.18
httpx==0.24.1
gTTS==2.5.0
